- Provide place holder queries for Dao - Convert the count value from int to long - Retrofit the interface and its implementation for int to long conversion

+++ src/main/java/org/gooru/nucleus/handlers/classes/processors/repositories/activejdbc/dbhandlers/classactivities/activitylist/common/contentfetcher/ActivityFetcher.java
@@ -15,7 +15,7 @@ public interface ActivityFetcher {
 
   List<AJEntityClassContents> fetchContents();
 
-  int fetchTotalContentCount();
+  Long fetchTotalContentCount();
 
   static ActivityFetcher buildContentFetcherForOnlineScheduledActivities(
       ListOnlineScheduledActivityCommand command) {
