+++ src/main/java/org/gooru/nucleus/handlers/classes/processors/repositories/activejdbc/dbhandlers/classactivities/activitylist/common/contentfetcher/ContentFetcherForUnscheduledActivities.java
@@ -31,10 +31,10 @@ class ContentFetcherForUnscheduledActivities implements ActivityFetcher {
   }
 
   @Override
-  public int fetchTotalContentCount() {
+  public Long fetchTotalContentCount() {
     if (!contentFetchDone) {
       throw new IllegalStateException("Count fetch without fetching content");
     }
-    return contents != null ? contents.size() : 0;
+    return contents != null ? (long) contents.size() : 0;
   }
 }
